Culled unused pages
Removed the ethical reflections page as these will be submitted individually on PDF. The files are still accessible for reference.
Meeting minutes page removed due to not being used.
Marking rubric page removed as this did not contain useful information, the rubric is now on canvas.
Redesigned guideline conformance, and portfolio pages.